+++ src/server.js
@@ -10,17 +10,24 @@
 import 'babel-polyfill';
 import path from 'path';
 import express from 'express';
+import cookieParser from 'cookie-parser';
+import bodyParser from 'body-parser';
+import expressJwt from 'express-jwt';
+import expressGraphQL from 'express-graphql';
+import jwt from 'jsonwebtoken';
 import React from 'react';
 import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom/server';
 import { match, RouterContext } from 'react-router';
 import configureStore from './configureStore';
 import assets from './assets';
-import { port } from './config';
+import { port, auth } from './config';
 import makeRoutes from './routes';
 import ContextHolder from './core/ContextHolder';
 import Html from './components/Html';
 import Restaurant from './models/Restaurant';
+import passport from './core/passport';
+import schema from './data/schema';
 
 const server = global.server = express();
 
@@ -37,12 +44,46 @@ global.navigator.userAgent = global.navigator.userAgent || 'all';
 // Register Node.js middleware
 // -----------------------------------------------------------------------------
 server.use(express.static(path.join(__dirname, 'public')));
+server.use(cookieParser());
+server.use(bodyParser.urlencoded({ extended: true }));
+server.use(bodyParser.json());
+
+//
+// Authentication
+// -----------------------------------------------------------------------------
+server.use(expressJwt({
+  secret: auth.jwt.secret,
+  credentialsRequired: false,
+  /* jscs:disable requireCamelCaseOrUpperCaseIdentifiers */
+  getToken: req => req.cookies.id_token,
+  /* jscs:enable requireCamelCaseOrUpperCaseIdentifiers */
+}));
+server.use(passport.initialize());
+
+server.get('/login/facebook',
+  passport.authenticate('facebook', { scope: ['email', 'user_location'], session: false })
+);
+server.get('/login/facebook/return',
+  passport.authenticate('facebook', { failureRedirect: '/login', session: false }),
+  (req, res) => {
+    const expiresIn = 60 * 60 * 24 * 180; // 180 days
+    const token = jwt.sign(req.user, auth.jwt.secret, { expiresIn });
+    res.cookie('id_token', token, { maxAge: 1000 * expiresIn, httpOnly: true });
+    res.redirect('/');
+  }
+);
 
 //
 // Register API middleware
 // -----------------------------------------------------------------------------
 server.use('/api/content', require('./api/content').default);
 server.use('/api/restaurants', require('./api/restaurants').default);
+server.use('/graphql', expressGraphQL(req => ({
+  schema,
+  graphiql: true,
+  rootValue: { request: req },
+  pretty: process.env.NODE_ENV !== 'production',
+})));
 
 //
 // Register server-side rendering middleware